move PanelViewControl to HasViewControl
