+++ src/UI/Component/Panel/Standard.php
@@ -3,10 +3,11 @@
 /* Copyright (c) 2016 Timon Amstutz <timon.amstutz@ilub.unibe.ch> Extended GPL, see docs/LICENSE */
 
 namespace ILIAS\UI\Component\Panel;
+use ILIAS\UI\Component\ViewControl\HasViewControls;
 
 /**
  * This describes a Standard Panel.
  */
-interface Standard extends Panel, PanelViewControls
+interface Standard extends Panel, HasViewControls
 {
 }
