+++ src/UI/Implementation/Component/Panel/Standard.php
@@ -4,6 +4,7 @@
 
 namespace ILIAS\UI\Implementation\Component\Panel;
 
+use ILIAS\UI\Implementation\Component\ViewControl\HasViewControls;
 use ILIAS\UI\Component as C;
 
 /**
@@ -12,5 +13,5 @@ use ILIAS\UI\Component as C;
  */
 class Standard extends Panel implements C\Panel\Standard
 {
-    use PanelViewControls;
+    use HasViewControls;
 }
